requirements.txt
torch
torchvision
Pillow
numpy
matplotlib
pytorch_msssim